requirements.txt
PyQt5
deep-translator
Pillow
easyocr
pynput
openai
google-generativeai
opencv-python
